refactor: Extract stat-related functionality from SubredditResources into its own class
